[LFXV2-1730] refactor: enforce domain/infrastructure/service layer structure
Move packages to align internal/ with a strict three-layer layout:

  internal/log/        → internal/infrastructure/log/
  internal/middleware/ → internal/infrastructure/middleware/
  internal/service/notifications_render.go + templates/
                       → internal/service/email/ (new package)

The email package exports RenderProjectRoleNotification and
ProjectRoleNotificationData; adding a new email type means adding one
file to internal/service/email/ alongside its templates.

Update CLAUDE.md:
- Reflect new directory tree
- Bump Go version to 1.24+
- Add LFX_SELF_SERVE_BASE_URL to env vars table
- Clarify the two distinct NATS patterns (request/reply RPC vs
  event subscriptions) and which file each belongs in
