+++ cmd/project-api/main.go
@@ -27,9 +27,9 @@ import (
 	"github.com/linuxfoundation/lfx-v2-project-service/internal/domain"
 	"github.com/linuxfoundation/lfx-v2-project-service/internal/domain/models"
 	"github.com/linuxfoundation/lfx-v2-project-service/internal/infrastructure/auth"
+	"github.com/linuxfoundation/lfx-v2-project-service/internal/infrastructure/log"
+	"github.com/linuxfoundation/lfx-v2-project-service/internal/infrastructure/middleware"
 	internalnats "github.com/linuxfoundation/lfx-v2-project-service/internal/infrastructure/nats"
-	"github.com/linuxfoundation/lfx-v2-project-service/internal/log"
-	"github.com/linuxfoundation/lfx-v2-project-service/internal/middleware"
 	"github.com/linuxfoundation/lfx-v2-project-service/internal/service"
 	"github.com/linuxfoundation/lfx-v2-project-service/pkg/constants"
 	"github.com/linuxfoundation/lfx-v2-project-service/pkg/utils"
