+++ internal/service/email/templates/project_role_notification.html
